refactor(auth): Move AuthMiddleware to the main module
